+++ packages/asset-swapper/src/utils/market_operation_utils/sampler_operations.ts
@@ -20,6 +20,7 @@ import {
 import {
     BALANCER_V2_VAULT_ADDRESS_BY_CHAIN,
     BANCOR_REGISTRY_BY_CHAIN_ID,
+    CLIPPER_INFO_BY_CHAIN,
     DODOV1_CONFIG_BY_CHAIN_ID,
     DODOV2_FACTORIES_BY_CHAIN_ID,
     KYBER_CONFIG_BY_CHAIN_ID,
@@ -359,9 +360,10 @@ export class SamplerOperations {
         takerToken: string,
         takerFillAmounts: BigNumber[],
         gasCost: number,
+        source: ERC20BridgeSource = ERC20BridgeSource.LiquidityProvider,
     ): SourceQuoteOperation<LiquidityProviderFillData> {
         return new SamplerContractOperation({
-            source: ERC20BridgeSource.LiquidityProvider,
+            source,
             fillData: {
                 poolAddress: providerAddress,
                 gasCost,
@@ -378,9 +380,10 @@ export class SamplerOperations {
         takerToken: string,
         makerFillAmounts: BigNumber[],
         gasCost: number,
+        source: ERC20BridgeSource = ERC20BridgeSource.LiquidityProvider,
     ): SourceQuoteOperation<LiquidityProviderFillData> {
         return new SamplerContractOperation({
-            source: ERC20BridgeSource.LiquidityProvider,
+            source,
             fillData: {
                 poolAddress: providerAddress,
                 gasCost,
@@ -1322,9 +1325,9 @@ export class SamplerOperations {
                                 takerToken,
                                 makerToken,
                             ) || []
-                        ).map(poolAddress =>
+                        ).map(balancerPool =>
                             this.getBalancerSellQuotes(
-                                poolAddress,
+                                balancerPool,
                                 makerToken,
                                 takerToken,
                                 takerFillAmounts,
@@ -1358,9 +1361,9 @@ export class SamplerOperations {
                                 takerToken,
                                 makerToken,
                             ) || []
-                        ).map(poolAddress =>
+                        ).map(creamPool =>
                             this.getBalancerSellQuotes(
-                                poolAddress,
+                                creamPool,
                                 makerToken,
                                 takerToken,
                                 takerFillAmounts,
@@ -1449,6 +1452,32 @@ export class SamplerOperations {
 
                         return this.getLidoSellQuotes(lidoInfo, makerToken, takerToken, takerFillAmounts);
                     }
+                    case ERC20BridgeSource.Clipper:
+                        const { poolAddress: clipperPoolAddress, tokens: clipperTokens } = CLIPPER_INFO_BY_CHAIN[
+                            this.chainId
+                        ];
+                        if (
+                            clipperPoolAddress === NULL_ADDRESS ||
+                            !clipperTokens.includes(makerToken) ||
+                            !clipperTokens.includes(takerToken)
+                        ) {
+                            return [];
+                        }
+                        // Clipper requires WETH to be represented as address(0)
+                        const adjustedMakerToken =
+                            makerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : makerToken;
+                        const adjustedTakerToken =
+                            takerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : takerToken;
+                        // Supports the PLP interface
+                        return this.getLiquidityProviderSellQuotes(
+                            clipperPoolAddress,
+                            adjustedMakerToken,
+                            adjustedTakerToken,
+                            takerFillAmounts,
+                            // tslint:disable-next-line: custom-no-magic-numbers
+                            0, // Not used for Clipper
+                            ERC20BridgeSource.Clipper,
+                        );
                     default:
                         throw new Error(`Unsupported sell sample source: ${source}`);
                 }
@@ -1718,6 +1747,32 @@ export class SamplerOperations {
 
                         return this.getLidoBuyQuotes(lidoInfo, makerToken, takerToken, makerFillAmounts);
                     }
+                    case ERC20BridgeSource.Clipper:
+                        const { poolAddress: clipperPoolAddress, tokens: clipperTokens } = CLIPPER_INFO_BY_CHAIN[
+                            this.chainId
+                        ];
+                        if (
+                            clipperPoolAddress === NULL_ADDRESS ||
+                            !clipperTokens.includes(makerToken) ||
+                            !clipperTokens.includes(takerToken)
+                        ) {
+                            return [];
+                        }
+                        // Clipper requires WETH to be represented as address(0)
+                        const adjustedMakerToken =
+                            makerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : makerToken;
+                        const adjustedTakerToken =
+                            takerToken === NATIVE_FEE_TOKEN_BY_CHAIN_ID[this.chainId] ? NULL_ADDRESS : takerToken;
+                        // Supports the PLP interface
+                        return this.getLiquidityProviderBuyQuotes(
+                            clipperPoolAddress,
+                            adjustedMakerToken,
+                            adjustedTakerToken,
+                            makerFillAmounts,
+                            // tslint:disable-next-line: custom-no-magic-numbers
+                            0, // Not used for Clipper
+                            ERC20BridgeSource.Clipper,
+                        );
                     default:
                         throw new Error(`Unsupported buy sample source: ${source}`);
                 }