+++ contracts/dev-utils/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v1.3.35 - _August 11, 2021_
+
+    * Dependencies updated
+
+## v1.3.34 - _August 6, 2021_
+
+    * Dependencies updated
+
 ## v1.3.33 - _June 22, 2021_
 
     * Dependencies updated