+++ contracts/exchange-forwarder/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v4.2.37 - _August 11, 2021_
+
+    * Dependencies updated
+
+## v4.2.36 - _August 6, 2021_
+
+    * Dependencies updated
+
 ## v4.2.35 - _June 22, 2021_
 
     * Dependencies updated