+++ contracts/treasury/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v1.3.1 - _August 11, 2021_
+
+    * Dependencies updated
+
+## v1.3.0 - _August 6, 2021_
+
+    * Added proposal 1 params and test (#298)
+
 ## v1.2.3 - _June 22, 2021_
 
     * Dependencies updated