+++ contracts/zero-ex/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v0.27.1 - _August 11, 2021_
+
+    * Dependencies updated
+
+## v0.27.0 - _August 6, 2021_
+
+    * Add `Clipper` as a custom liquidity source
+
 ## v0.26.0 - _June 22, 2021_
 
     * Add Lido stETH deposit integration (#260)